+++ CHANGELOG.md
@@ -10,3 +10,7 @@
   - FEATURE: `node._parent` returns the parent object
   - FEATURE: `for x in list._xpath_sorted` returns items sorted by xpath `for x in list` rerturns items based on the datastore order (which should be the order they were added)
   - Docker file split into `builder` for compiling everything from source and a smaller version (`docker/build.sh`) to build both images.
+- devel -
+  - FIX: return a PresenceContainer object when calling create on a presence node path.
+  - FIX: implement `has_item(xpath)` asking the datastore about existence rather than getting the full list back.
+  - Renamed BlackArt to Voodoo for representation of objects.